change register page to work with antd and update mongoose version
