+++ client/src/components/auth/register/index.js
@@ -1,16 +1,19 @@
 import React, { Component } from "react";
 import PropTypes from "prop-types";
+import { withRouter } from "react-router-dom";
 import { Form, Button, Row, Col } from "antd";
 import { connect } from "react-redux";
-import { loginUser } from "../../../actions/authActions";
+import { registerUser } from "../../../actions/authActions";
 import TextFieldGroup from "../../common/TextFieldGroup";
 
-class Login extends Component {
+class Register extends Component {
   constructor() {
     super();
     this.state = {
+      name: "",
       email: "",
       password: "",
+      password2: "",
       errors: {}
     };
   }
@@ -22,49 +25,55 @@ class Login extends Component {
   }
 
   componentWillReceiveProps(nextProps) {
-    if (nextProps.auth.isAuthenticated) {
-      this.props.history.push("/dashboard");
-    }
-
     if (nextProps.errors) {
       this.setState({ errors: nextProps.errors });
     }
   }
 
+  onChange = e => {
+    this.setState({ [e.target.name]: e.target.value });
+  };
+
   onSubmit = e => {
     e.preventDefault();
 
-    const userData = {
+    const newUser = {
+      name: this.state.name,
       email: this.state.email,
-      password: this.state.password
+      password: this.state.password,
+      password2: this.state.password2
     };
 
-    this.props.loginUser(userData);
-  };
-
-  onChange = e => {
-    this.setState({ [e.target.name]: e.target.value });
+    this.props.registerUser(newUser, this.props.history);
   };
 
   render() {
     const { errors } = this.state;
 
     return (
-      <div className="login">
+      <div className="register">
         <Row type="flex" justify="center">
           <Col span={24}>
-            <h1 className="display-41 text-center1">Log In</h1>
+            <h1 className="display-41 text-center1">Sign Up</h1>
             <p className="lead1 text-center1">
-              Sign in to your DevConnector account
+              Create your DevConnector account
             </p>
             <Form noValidate onSubmit={this.onSubmit}>
               <TextFieldGroup
-                placeholder="Email Address"
+                placeholder="Name"
+                name="name"
+                value={this.state.name}
+                onChange={this.onChange}
+                error={errors.name}
+              />
+              <TextFieldGroup
+                placeholder="Email"
                 name="email"
                 type="email"
                 value={this.state.email}
                 onChange={this.onChange}
                 error={errors.email}
+                info="This site uses Gravatar so if you want a profile image, use a Gravatar email"
               />
               <TextFieldGroup
                 placeholder="Password"
@@ -73,6 +82,14 @@ class Login extends Component {
                 value={this.state.password}
                 onChange={this.onChange}
                 error={errors.password}
+              />
+              <TextFieldGroup
+                placeholder="Confirm Password"
+                name="password2"
+                type="password"
+                value={this.state.password2}
+                onChange={this.onChange}
+                error={errors.password2}
               />
               <Button
                 size="large"
@@ -90,8 +107,8 @@ class Login extends Component {
   }
 }
 
-Login.propTypes = {
-  loginUser: PropTypes.func.isRequired,
+Register.propTypes = {
+  registerUser: PropTypes.func.isRequired,
   auth: PropTypes.object.isRequired,
   errors: PropTypes.object.isRequired
 };
@@ -101,9 +118,7 @@ const mapStateToProps = state => ({
   errors: state.errors
 });
 
-const WrappedLogin = Form.create()(Login);
-
 export default connect(
   mapStateToProps,
-  { loginUser }
-)(WrappedLogin);
+  { registerUser }
+)(withRouter(Register));
