+++ package.json
@@ -20,7 +20,7 @@
     "express": "^4.16.3",
     "gravatar": "^1.6.0",
     "jsonwebtoken": "^8.2.0",
-    "mongoose": "^5.0.12",
+    "mongoose": "^5.3.8",
     "passport": "^0.4.0",
     "passport-jwt": "^4.0.0",
     "validator": "^9.4.1"
